test: Fix test code
